feat: replace Vercel Functions with GitHub Actions
